Create Query and Insert
Updated application so it is more secure. Eliminated SQL concatenated INSERT commands for security reasons.

Uploaded and will be working with the "Adventure Works" Database for learning purposes.

The first tab of the application allows searches into the SQL Adventure Works database. The second tab allows the user to add work shifts to the humanresources.shifts table in the database.  

***The text box at the bottom was used for testing code to see what the output looked like and ensure it fits with SQL field type values